requirements.txt
torch>=2.0.0
transformers>=4.30.0
lightning>=2.0.0
sentence-transformers>=3.3.1
timm>=0.9.0
pillow>=9.0.0
numpy>=1.24.0
deepspeed>=0.12.0
nltk>=3.8
boto3>=1.26.0
sacremoses>=0.0.53
tensorboard>=2.13.0
jupyterlab>=4.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
jsonargparse[signatures]>=4.20.0
spacy>=3.6.0
pandas>=2.0.0
datasets>=2.14.0
accelerate>=0.21.0
wandb>=0.15.0
torchvision>=0.15.0
einops>=0.7.0
